Complete Rough Draft of Library Class
This is broken. Employees will end up reading the same book more than once.  I'm going to try to fix this by adding a list of books read to each employee. Then when a book is being passed it will check to see if the employee has already read the book. Iterating over the PQ until it finds one that has and then follows through with the rest of the function.